+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Store computetasks logs address and owner in localrep (#801)
+
 ## [0.9.0] 2022-03-01
 
 ### Added